+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/mailer_multiple_transport.php
@@ -1,0 +1,11 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'mailer' => [
+        'transports' => [
+            'baz' => 'smtp://baz',
+            'foo' => 'smtp://foo',
+            'bar' => 'smtp://bar',
+        ],
+    ],
+]);